# Optimize frame encodeValue unencodeValue, replace with more efficient version