--- a/frame/encode.go
+++ b/frame/encode.go
@@ -1,27 +1,36 @@
 package frame
 
 import (
+	"bytes"
 	"strings"
 )
 
+var (
+	replacerForEncodeValue = strings.NewReplacer(
+		"\\", "\\\\",
+		"\r", "\\r",
+		"\n", "\\n",
+		":", "\\c",
+	)
+	replacerForUnencodeValue = strings.NewReplacer(
+		"\\r", "\r",
+		"\\n", "\n",
+		"\\c", ":",
+		"\\\\", "\\",
+	)
+)
+
 // Encodes a header value using STOMP value encoding
-// TODO: replace with more efficient version.
 func encodeValue(s string) []byte {
-	s = strings.Replace(s, "\\", "\\\\", -1)
-	s = strings.Replace(s, "\r", "\\r", -1)
-	s = strings.Replace(s, "\n", "\\n", -1)
-	s = strings.Replace(s, ":", "\\c", -1)
-	return []byte(s)
+	var buf bytes.Buffer
+	buf.Grow(len(s))
+	replacerForEncodeValue.WriteString(&buf, s)
+	return buf.Bytes()
 }
 
 // Unencodes a header value using STOMP value encoding
-// TODO: replace with more efficient version.
 // TODO: return error if invalid sequences found (eg "\t")
 func unencodeValue(b []byte) (string, error) {
-	s := string(b)
-	s = strings.Replace(s, "\\r", "\r", -1)
-	s = strings.Replace(s, "\\n", "\n", -1)
-	s = strings.Replace(s, "\\c", ":", -1)
-	s = strings.Replace(s, "\\\\", "\\", -1)
+	s := replacerForUnencodeValue.Replace(string(b))
 	return s, nil
 }