@hanjm 2019-02-10
you can see the std lib code, strings/replace.go https://golang.org/src/strings/replace.go#L315:

```
func (r *genericReplacer) Replace(s string) string {
	buf := make(appendSliceWriter, 0, len(s))
	r.WriteString(&buf, s)
	return string(buf)
}
```
using `[]byte(replacerForEncodeValue.Replace(s))` make unnecessary bytes to string and strings to bytes transformation.